Combining objects_run3_2lep.py and objects_wwz.py
